Add LICENSE to the code
Dear all, this is an attempt to license the code of the project. I am proposing this license, as a good option that is liberal and allows sharing without any limitations.

For this to happen, we need the okay from every contributor, otherwise we cannot do this, and may have to replace their contributions.

All contributors of code or documents have been added to the list, ordered by the number of commits, descending, and with the copyright starting from the first year of their contribution. Please make sure the names are okay.